renderer_vulkan: Abstract descriptor management
* The pipeline cache was starting to get cluttered
